Revert "[WIP] testing backend submission (#11)"
This reverts commit e90e771fbc22c6fe629513e3003166e29555722b.
