+++ bin/test_submission
@@ -1,0 +1,2 @@
+#!/usr/bin/env python
+from rampwf import test_submission  # noqa
