+++ rampwf/__init__.py
@@ -2,6 +2,4 @@ from . import prediction_types
 from . import score_types
 from . import workflows
 
-__all__ = ['kits', 'utils']
-
 __version__ = '0.1.dev'
