--- rampwf/kits/__init__.py
@@ -1,4 +1,0 @@
-from .ramp_kit import fetch_ramp_kit
-
-
-__all__ = ['fetch_ramp_kit']
